feat: use P2P fields in batch form
